Move Bridge pool VP init logic to storage mod
