+++ ethereum_bridge/src/lib.rs
@@ -1,6 +1,5 @@
 extern crate core;
 
-pub mod bridge_pool_vp;
 pub mod oracle;
 pub mod parameters;
 pub mod protocol;
