+++ ethereum_bridge/src/storage/mod.rs
@@ -1,4 +1,5 @@
 //! Functionality for accessing the storage subspace
+pub mod bridge_pool_vp;
 pub mod eth_bridge_queries;
 pub mod proof;
 pub mod vote_tallies;
